requirements.txt
discord.py==2.1.0
PyNaCl==1.5.0
asyncpg==0.27.0
wavelink==1.3.3